+++ src/main/java/com/cognite/client/Relationships.java
@@ -89,12 +89,6 @@ public abstract class Relationships extends ApiBase {
         List<String> partitions = buildPartitionsList(getClient().getClientConfig().getNoListPartitions());
 
         return this.list(requestParameters, partitions.toArray(new String[partitions.size()]));
-
-
-        // The relationships API endpoint does not support partitions (yet). Therefore no partitions are used here
-        // in the list implementation. Convert to using partitions when the relationship endpoint is updated.
-
-        //return list(requestParameters, new String[0]);
     }
 
     /**